+++ src/page/archiving/index/component/TimeBlockModal/component/Block/Date/BlockDate.tsx
@@ -1,5 +1,8 @@
-import { supportStyle, textStyle } from '@/page/archiving/timeBlockModal/component/Block/Date/BlockDate.style';
-import useDateRange from '@/page/archiving/timeBlockModal/hook/common/useDateRange';
+import {
+  supportStyle,
+  textStyle,
+} from '@/page/archiving/index/component/TimeBlockModal/component/Block/Date/BlockDate.style';
+import useDateRange from '@/page/archiving/index/component/TimeBlockModal/hook/common/useDateRange';
 import { css } from '@emotion/react';
 
 import Flex from '@/common/component/Flex/Flex';
