+++ src/page/archiving/index/component/TimeBlockModal/component/Upload/File/Add/BlockAdd.tsx
@@ -1,5 +1,8 @@
-import { boxStyle, buttonStyle } from '@/page/archiving/timeBlockModal/component/Upload/File/Add/BlockAdd.style';
-import useFile from '@/page/archiving/timeBlockModal/hook/common/useFile';
+import {
+  boxStyle,
+  buttonStyle,
+} from '@/page/archiving/index/component/TimeBlockModal/component/Upload/File/Add/BlockAdd.style';
+import useFile from '@/page/archiving/index/component/TimeBlockModal/hook/common/useFile';
 
 import { Dispatch, SetStateAction } from 'react';
