+++ src/page/archiving/index/component/TimeBlockModal/component/Upload/File/List/BlockItem.tsx
@@ -1,4 +1,7 @@
-import { borderStyle, textStyle } from '@/page/archiving/timeBlockModal/component/Upload/File/List/BlockItem.style';
+import {
+  borderStyle,
+  textStyle,
+} from '@/page/archiving/index/component/TimeBlockModal/component/Upload/File/List/BlockItem.style';
 
 import Delete from '@/common/asset/svg/file-delete.svg?react';
 import Flex from '@/common/component/Flex/Flex';
