+++ src/shared/component/workSpaceModal/info/WorkSpaceInfo.tsx
@@ -2,7 +2,7 @@ import Flex from '@/common/component/Flex/Flex';
 import Heading from '@/common/component/Heading/Heading';
 import Text from '@/common/component/Text/Text';
 
-import { headingStyle, textStyle, topStyle } from '@/shared/component/workSpaceModal/info/WorkSpaceInfo.style';
+import { headingStyle, textStyle, topStyle } from '@/shared/component/WorkSpaceModal/info/WorkSpaceInfo.style';
 import { STEPS, STEPS_BY_CATEGORY } from '@/shared/constant';
 
 interface WorkSpaceInfoProps {
