+++ src/story/page/archiving/modal/BlockItem.stories.tsx
@@ -1,4 +1,4 @@
-import BlockItem from '@/page/archiving/timeBlockModal/component/Upload/File/List/BlockItem';
+import BlockItem from '@/page/archiving/index/component/TimeBlockModal/component/Upload/File/List/BlockItem';
 import type { Meta, StoryObj } from '@storybook/react';
 
 const meta = {
