+++ src/story/shared/modal/WorkSpaceImage.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import WorkSpaceImage from '@/shared/component/workSpaceModal/image/WorkSpaceImage';
-import { WorkSpaceProvider } from '@/shared/hook/common/useWorkSpaceContext';
+import { WorkSpaceProvider } from '@/shared/component/Modal/hook/useWorkSpaceContext';
+import WorkSpaceImage from '@/shared/component/WorkSpaceModal/image/WorkSpaceImage';
 
 const meta = {
   title: 'page/Modal/WorkSpaceImage',
